[sync] Add sync_cnd_t based on a generic implementation
Now that we have sync_mtx_t, we should also provide an implementation
of a condition variable that is compatible with that mutex.

The algorithm is borrowed from our fork of MUSL with minimal
modifications.  A generic implementation of the algorithm and a generic
test suite are provided and specialized for sync_mtx_t/sync_cnd_t.
The plan is to use the same template for c11 & pthread condvars to avoid
code duplication.

ZX-2803

TEST=Unit tests included

Change-Id: Iaaa435e5073703d03ccba0a991ebddd38252aeac