[zircon][oom] Better way to kill
Earlier this year the "job importance" system was removed
as it was not used and made things complex for job management.
However, the OOM lowmem handler was not modified enough.

Before this CL the lowmem handler would kill the process which
happened to be enumerated first and which was alive regardless
of its memory contribution. This effectively kills one of the
most newest processes which usually have little impact in
memory and effectively is 'random'.

In this CL we introduce ZX_PROP_JOB_KILL_ON_OOM which usermode
can use to mark the best processes to be terminated in out of
memory situations. This is by no means the final system but
we are informed that in several important cases there is a
particular job tree which should be terminated which today
has 3 properties:

  1- Gives a clear signal to the tester/user the system is dead
  2- Does not produce effects which can be easily construed as
       issues with particular pieces (e.g wireless is dead)
  3- Preserves logs across the network

In other words, in this CL we are probably in the right track.
The true heuristics are too subtle to be computable by Zircon
alone.

ZX-2731 #comment progress

TEST= use "k oom lowmem" to see the process in action. Requires
at least one job to be marked as ZX_PROP_JOB_KILL_ON_OOM via
zx_object_set_property().

Change-Id: I40d24027b1d1b90b6e8aa1a513095ed82463d93b