[fbl] Prepare for intrusive container refactor.
In a series of CLs, fbl instrusive containers will be refactored so
that they no longer hold any managed pointer in their internal node
state.  This will help static analysis tools to determine that we will
never accidentally trigger a recursive destruction of an entire
conainer's contents because a managed reference in a node state
accidenally got released.  See ZX-2529 for details.

As part of this refactor, much of the requirements of the
ContainerPtrTraits class will go away.  Pointer traits no longer need
to know how to do things like manipulate sentinel bits for managed
pointer types, or to swap managed pointer types, as all internal
pointers in the containers will be raw pointers.  All pointer traits
really need to know how to do is to leak a pointer reference into the
unmanaged world, and to reclaim that unmanaged pointer into a managed
reference as the pointer enters and leaves the container.

To prepare for this, begin by implementing some of this functionality
in the fbl::internal namespace inside container_utils.h

ZX-2529 #comment Prep for refactor.

Tests: Build and new unit tests
Change-Id: I898a58e109a6f64f22db589ead81a26decb5c43b