[scenic] Add SetDebugName to Session
Add |SetDebugName| to |Session|, which allows for sessions to set a
debug name.  The debug name can be used to make logging and tracing
easier to process (by both humans and scripts). For its first use, we
output the session debug name as an argument of the
"Session missed frame" event, which makes it much easier to figure out
which session dropped a frame when (e.g.) viewing a trace.

Test:
$ fx shell trace record --duration=1 &&
    fx shell cat /data/trace.json > trace.json &&
    cat trace.json | jq . | grep session_debug_name

Change-Id: I06d3567c16dae751a1158b6635540fe0d1926e0e