[wlan][minstrel] Fix a bug in BetterProbability
if max_probability's probability is higher than threshold,
a new candidate with lower throughput but higher probability will
replace it based on current implementation due to the second condition,
lhs.probability > rhs.probability.

This is not what we wanted. This change will only allow a candidate to
replace existing max_probability if max_probability's probability is
below threshold.

Test: Build and smoke. Verified behavior in local build.
NET-1327 #comment

Change-Id: Ie8e09f0d3ec3be5003afce8510eee8e6d1559976