[zxs] Add a low-level socket library
This library provides an abstraction over the zx::socket control plane
used to communicate with netstack and over the address framing used in
datagram sockets.

Eventually, FDIO will use this library to implement socket file
descriptors.

This CL removes the [Layout = "Simple"] annotation from Connectivity
because the FIDL compiler was generating an empty ops struct, which does
not compile with -Wextern-c-compat.

Test: No behavior yet.
ZX-1929 #comment

Change-Id: I7121f97c24811ada2a45dab474411e8f19b74cb8