[debugger] Resolve class members in base classes.
Previously printing values did not find members in base classes.

This change adds a standard way to iterate through a class hierarchy and
uses this to resolve class member names.

This also adds support for the "object pointer" in a class member function.
The object pointer is not currently used, but will be needed in the future
to resolve variables relative to "this".

TEST=unit

Change-Id: I4df7213e00b2f8479f4ee28e31a1434df2bb4a63