[debugger] Print base class values.
When printing classes and structs, this adds base class values to the
output.

Some refactoring of the resolve_member file was done to allow extraction
of base classes (very similar to member accesses). The file is also
renamed to resolve_collection.cc since this is where all Collection
resolution stuff goes.

TEST=manual,unit

Change-Id: Idc7ed9b2ac330dc2eed7e007b01d492a0bf693a6