[zircon][usb-request]Add protocol to get context sizes
usb-request_t structure has various fields that are public some of which
need to be restricted for usage in respective layers of usb stack(node,
complete_cb etc). For this reason usb_request_t will be broken into public
portion and private portion for each layer in the stack. This
changeset defines the private portions and adds protocol ops for getting the
sizes of the private structs. The topmost layer that wants to create the request
will alloc the size depending on the return values of the protocol ops.
More changesets which actually put these internal structs to use will follow.
After all the usage has been switched to the private structs, I will
remove the unnecessary fields from usb_request_t.

Test: Built zircon

Change-Id: I8a5f680f5b184e5ab7f307026027e99447b00bfe