[digest] Fix the move constructor for Digest
The internal "ref_count_" member of digest lacked a default
initializer, and the move contructor for Digest was not explicitly
initalizing it to zero, meaning that the value taken on by ref_count_
was non-deterministic.

This started to trigger ZX_DEBUG_ASSERTs in the Digest implementation
when DEBUG_ASSERTS were enabled in host tools (such as the BlobFS
builder) once debug assert was enabled in debug tools build with
I6b7365e795a55be3b5f50add8551cbac592238db.

Fix this issue by creating a default initializer the ref_count_ field
which gets shared by all constructors, including the move constructor
and any new constructors which might come along at a later date.

In addition, the move semantics for this class seemed to be a bit off.
In particular, they were failing to move the ctx_ member which was a
move only unique_ptr type.  Since this seems to be the primary reason
one would ever want to have a move constructor for this class at all,
it seems important to be clear in their definition.

Define the move semantics for this class to be the following...
++ Moving digest1 into digest2 moves any context in the process.
++ Moving digest1 into digest2 also copies any cached digest bytes
   from 1 into 2, and zeros out any cached digest bytes in 1,
   returning it to the default constructed state.

Finally, add a unit test to verify the new definition of move
constructon for the Digest.

Tests: new unit tests.
Change-Id: I16eb412573bc6a0b3771659ab733c17c57e25ccc