[debugger] Added extended flag for registers.
Moved x86 shift/mask constants to the debug_ipc header for exported use and
created better accessing macros.
The testing uses this to make them less criptic about what is testing
and better verify correct behaviour.

Porting the debug_agent to use this information will be done in a future
CL.

For x86 HW debug registers, we need to read some system flags of RFLAGS
that are not normally shown.

This CL adds a little refactoring by passing a context struct reference to
the filtering/formatting functions to modify their behaviour.

TEST=unit, manual

Change-Id: Id296a2455c248a5a510227a9e46b4e532e48c265