[wlan][rsn] Fix replay problem with msg1
When the Authenticator sends the first message of the Handshake,
and doesn't receive a response in time, it will replay the first
message, which is a totally valid behavior.
However, if the Supplicant sends a response to the first message,
before the replayed one arrives, it will move into a state in
which it waits for the third message of the handshake. Now that it
receives a replayed first message, it will reset the entire
handshake. In the meantime, the Authenticator received the response
to the handshake and responds with the third message of the handshake,
which is unexpected as the Supplicant never responded to the last
first message it received.

Test=Included regression test

NET-1705 #done

Change-Id: I252b5b89f7f4f9ebb9b2bef1a7ae5db2d4890b36