[auth] Add AuthenticationUIContext to Authorize
This supports setup from the user shell. If supplied, this overrides
the AuthenticationContextProvider provided at construction.

(Note the CL is downscoped from earlier versions that also included
expiry time, that is a much bigger change and its not completely clear
that it was the right answer vs redefining and reusing the auth provider
struct.

AUTH-110

Tested: CQ, no new functionality introduced yet.
Change-Id: I66705a21abcbeb1d01c572077c7ec8e9dadd5653