[kernel] Implementation of a blocking reader-writer lock
This is an initial implementation of a read-writer lock that blocks on a wait_queue
under contention. The lock acquisition relies on shared integer state variable that
has a fastpath acquisition involving a single atomic operation, with a more complicated
spin-lock fallback under contention.

Some choices in the lock still require benchmarking to inform whether they are correct
or not. Namely the use of weak compare_exchange and lack of preempt disable to remove
a potential race.

Tests have been done offline and most of the design validated as a spin model. Proper
unit tests and perf tests in upcoming commits.

ZX-3085 #comment

Test: compiles. Tested manually offline.

Change-Id: I041c941a9b58cba32fd2b033c1acd09da23fde75