[symbolize] Use ids.txt for all of "threads"
Since 1b2ef7187e7eef6de9ce384c014962ce573dc836, only one stack crawl
from the output of "threads" got the benefit of ids.txt, which caused us
to try to use stripped .so files.  With this, we use unstripped .so
files for all the stack crawls in the output of "threads", which makes
those stack crawls work better.

We used stripped .so files in particular because while exe.unstripped
is prepended to the search path in the fallback case, the
shared/lib.unstrippped dir is not prepended (which remains the case
after this commit).

Tested: threads, fx symbolize, force crash devhost, fx symbolize, CQ
Change-Id: Ifc807017a355c1f7527a0e0f64a99429d5779c3c