[images] Add blob.blk and data.blk to images.json
This adds blob.blk and data.blk to images.json, which allows
the tool make-fuchsia-vol to switch from image_paths.sh to
images.json. This was done because make-fuchsia-vol cannot
consume the fvm.blk images, because we can only use fvm.blk
if we size the partition to exactly fit it, but that can be
difficult because GPT requires padding by spec, which fvm
does not.

BLD-301 #done

Change-Id: I640ea5a60060e2cf5d1631aee7d7b0f37a3d754b