[bt][data] templatize SocketFactory
Abstract SocketFactory from a concrete class, to a class
template. This paves the way for RFCOMM sockets, which will
also use this template.

L2CAP sockets are supported by providing an explicit instantiation of
the SocketFactory, with the apropriate type parameters. A header file
provides an alias for the template instance, for ease of use.

Also:
* Move SocketFactory to the internal sub-namespace of btlib::data,
  since SocketFactory should not be referenced outside of btlib::data.
* Make SocketFactory final, since it is not designed to be subclassed.

Bug: NET-1718
Test: bt-host-unittests (on qemu)

Change-Id: I278fd6ac65ff885c63b5c601378964bc22a34ce2