[bt][data] templatize SocketChannelRelay
Abstract SocketChannelRelay from a concrete class, to a class
template. This paves the way for RFCOMM sockets, which will
instantiate the SocketChannelRelay template with RFCOMM-specific type
parameters.

L2CAP sockets are supported by providing an explicit instantiation of
the SocketChannelRelay, with the apropriate type parameters. A header
file is provided with an alias for the template instance, for ease of
use.

Also: Use C++17 nested namespace declarations. These are more compact
than the alternative, and will also ease migration to flat namespaces,
if/when we choose to do so. (For the argument against nested namespaces,
see https://fuchsia.googlesource.com/docs/+/master/development/languages/c-cpp/cpp-style.md#namespace-names)

Bug: NET-1718
Test: fx run-test bt-host-unittests (on eve)

Change-Id: Idd82adaad84068c114f6aa61c9123a9244a58ca9