[guest] Add new package for biscotti_guest.
Sets up the base guest package and scripts to build the kernel and an
empty stateful partition (which is not at all stateful with the current
implementation). No root disk image is provided yet, but it's expected
you provide on in //garnet/bin/guest/pkg/biscotti_guest/images/$ARCH/disk.img.

Test: guest launch biscotti_guest
Change-Id: I3e94935d31bd97d28d62b4dbffc805f6fa313941