[minfs] Add new functionality to AllocatorPromise.
Allow an AllocatorPromise to split its remaining reserved count
with another AllocatorPromise. Also allow an AllocatorPromise to
exist in an "uninitialized" state (i.e., not yet tied to any
Allocator or reservation).

Additionally, add an AllocatorStorage object to the Allocator
which implements persistence of allocation data to disk. This
allows the Allocator to be more easily tested by using a
FakeStorage instance.

Test: allocation_tests

Change-Id: I082665d06a315535c34d4c0e16f045c3d6f5a077