[fbl] Simplify ContainerPtrTraits
Now that the refactor is complete, and fbl intrusive containers deal
pretty much exclusively with raw pointers internally, we can simplify
the requirements for the various pointer trait classes which allow us
to support various managed/unmanaged pointer types.

Finish the cleanup by removing all of the methods from the pointer
traits which are no longer needed.  This includes Take, Swap, and all
of the things related to container sentinels (which have been moved to
the internal container pointer utils)

Tests: Build only
Change-Id: Ifc6cb282f1e4c95a19f72bfa68558506163a6685