[mediaplayer] expose can_pause and can_seek
This CL plumbs can_pause and can_seek up to clients. can_pause is currently
always true because we don't support any non-pausing source at the moment, but
this will change when we implement bluetooth A2DP rendering.

The mediaplayer FIDL service logs a warning and otherwise ignores pause when
can_pause if false. Same for seek when can_seek is false. Lower layers DCHECK
that the capabilities are being honored.

TEST: manual: played a video, checked introspection
Change-Id: I9d0877c8f6f42491394aec4c6328a8c2b2faeeb5