[ledger] Update PageUsageListener API.
The current PageUsageListener API is problematic for our current use
case. When a page closes we want to mark it as closed and try to evict
it if empty. We can and should only mark it as closed when external
requests to a page are all closed. However, we can't evict it unless
both internal and extrernal requests are finished.

This CL updates the API to make this distinction.

It also fixes a bug on the OnPageClosed being called more times than it
should: It was called every time the on_empty_callback from PageManager
was called. However, that could be called multiple times even when the
previous state was still empty. To solve this, in the CL we check that
the page is currently considered open before sending the OnPageClosed
notification.

Test: DiskCleanupManagerTest, DontEvictPagesOnPageClosed
      LedgerManagerTest, OnPageOpenedClosedUnused
Change-Id: Ia1d8833bb1b6e91f31f274f76d187a12eb33b51c