[mediaplayer] Add cache options to demux.
Adds cache options to demux in terms of time. Demux is now constructed
with a ReaderCache instead of Reader to be aware of those settings.

chunk_size is removed as a cache option for ReaderCache because no class
above it will have the information needed to appropriately set it. A
future CL will decide that based on timing observations ReaderCache
makes internally.

TEST: mediaplayer_tests

Change-Id: Ibdbc13a02d4379516380405ed4c845488187129d