[debugger] Implicitly resolve varables on |this|
Unqualified names are now resolved on |this| if not found as variables
or parameters.

Adds some additional testing infrastructure to help creating
collections. The old "MakeStructTwoMembers" is replaced with a more
flexible variant that takes a list of members, and also allows those
members to be offset to allow for precice placement of derived classes.

The memory requesting system is enhanced to more flexibly return data
inside subranges. A test for this mock is added.

TEST=unit,manual

Change-Id: Ibbb977431e5bb69f43eb4cbf9acd951182819e31