[loader] Lay groundwork for package-updating loader
- Subclass RootLoader from ComponentLoader, which implements most of
the logic. ComponentLoader lives in lib/loader and can be subclassed
from other places.
- Standardize on FuchsiaPkgUrl as the internal way to identify a
component in a package. Specifically, pass this instead of a file:// URL
when naming the default component in a package

ComponentLoader will later be used as a superclass of
PackageUpdatingLoader, for CP-169. This allows us to factor out common
code for resolving package URLs.

CP-169 #progress

TESTED=new tests for fuchsia_pkg_url and cmx

Change-Id: I8bb560021752b073d440f36029bbd8518c9dc2b4