[story_shell] Signal defering of gestures to Session Shell
In the absence of gesture disambiguation, StoryShell needs to signal to
Session Shell whether or not it needs to handle back gestures
(This allows us to use the same gesture to dismiss Surfaces and
Stories as appropriate).

When Story Shell has no dismissible Surfaces in the Story,
it signals it is not handling back gestures, and Session Shell can use
that gesture to dismiss the story. Otherwise it signals it is handling,
and expects Session Shell to defer the gesture.

test: manual
Change-Id: I0f687c49c1b38f9cb963f5fc6d726fa78b2b17c9