[zircon][sdmmc] Do not create worker event for sdio
worker event is not used for sdio. We are currently submitting all the
sdio transactions synchronously on the client's thread itself. This
changeset moves the worker event to be created only for mmc device.

Test: Verified on astro that there is no crash when calling
sdmmc_release. Verified on vim2 that the mmc comes up and that I can
use dd.

Change-Id: I713b1beabf7f74a58abfb1d33883d419d6850ca7