[sysmem] Temporarily work around union alignment rules in C bindings
The fix for FIDL-204 is in flight, but other work is blocked behind
this FIDL interface being usable for the C bindings. Since all the
unions involved have a single member, make the types structs for
now. I will make them unions again as soon as FIDL-204 is resolved.

FIDL-204 #comment

Test: cq
Change-Id: I19ea90f2af671f1e1a07097a600e5ae07c95cca8