[pkg] Add --auto_update_packages build flag
If garnet is built with this flag, packaged components will be automatically
updated before they are launched (or if the package is not on the device
in the first place, it will be installed).

This change is a hack of sysmgr and is not intended to be a production
feature. Instead, it is a special mode to allow the Modular team to
experiment with ephemeral mods and automatic package updates. It can
also be used to speed up development cycles, by eliminating the step of
restarting the device or pushing a package.

Example sequence of commands (host-side) to run a component with
auto-updates:

$ fx set x64 --args auto_update_packages=true
$ fx full-build
...
$ fx serve-updates&
$ fx run -N  # if using qemu
...
$ fx shell 'run my_component'  # will launch component in sys realm

Components must be started in the sys realm or below to receive this
feature. The feature should also work if components are started from the
devmgr shell (the shell started in qemu), since that gets sysmgr's services.

CP-169 #comment

TESTED=package_updating_loader integration test. Also manually started a
component and looked for update

Change-Id: I79aa7972c65d045b0dd1f4b2c8878b1bcd077396