[garnet] use connectivity-testing crate
Now that the wlan_service_util file is in the connectivity-testing
crate, we can go ahead and remove this copy and use the crate instead.

In addition to running the smoke test, also verified the
connectivity-testing unit tests are working and that the smoke test unit
tests are no longer found.

BUG: NET-1646
Test: fx run-test connectivity_testing_lib_test_rustc (all passed)
Test: fx run-test wlan_smoke_test_bin_test_rustc (unknown target - as
expected)

Change-Id: Ie63a9153cc0a0f398459d5ba03e263ef7d1fb75f