[biscotti][guest] Automate container setup using gRPC.
This adds a CLI app called biscotti that starts a `biscotti_guest` and
uses gRPC over vsock to configure a debian container. To get networking
functional in the guest, the host netstack needs to be killed before
starting biscotti. Additionally, the network parameters for the guest
are hard-coded, so this will need to be updated to match the local
network configuration to be functional.

Test: $ run biscotti
      ....
      $ guest serial 0 3
      machina@stretch:~$
Change-Id: I757d22c0aed1192c1d54b00a65e480524d5d0941