[Scenic] Add Render Event to View/ViewHolder
Create an event that can be signaled when a View is rendering contents
to the screen. The event is currently signaled directly on the View. In
the future, pass this event to be signalled with the release fences when
the View is presented and its contents have been rendered with a Scenic
frame.

SCN-944 #comment

Tested: added and ran view_unittests

Change-Id: Id7b12c7988a4fe2b26df5de9f5b1bd242e8af239