[images] Have pack-images actions explictly depend on zedboot images
pack-images actions are fed zedboot_images.json, but they do not
currently depend explicitly on any zedboot image deps. Rather, they
depend on a default-images target, which in turn coincidentally depends
on _most_ of the zedboot image dependencies. The dependency it misses is
//build/images/zeboot:signed, which doesn't normally exist; when it does,
however, trying to build a build-archive will fail.

Change-Id: I1c5891d0b78ad9adf602718ae2a951fcbc00bcac