[set-device] introduce a default-device concept
Common use:

  fx set x64
  fx set-device my-x64-device-name

  fx set arm64
  fx set-device my-arm64-device-name

After the above, when `fx use out/x64` the default device for `fx serve` and
other such scripts will be "my-x64-device-name", and when `fx use out/arm64`,
it will be "my-arm64-device-name".

Users can override the default device for a single command invocation using
the global fx flag `-d`:
  fx -d <device-name> <command>

The set-device command persists to a file that lives alongside the build dir,
specifically `${FUCHSIA_BUILD_DIR}.device`. It is kept outside the build dir
as a convenience in order to not be clobbered by `gn clean`.

Test: Run `fx -x boot`, observe no device is specified (-a flag)
Test: Run `fx set-device`, observe pair is set
Test: Run `fx -x boot`, observe pair is passed to -a.
Test: Run `fx -d my-device-name -x boot` observe device is passed to -a.
Test: Run `fx -x serve` observe -a also has device name.
Test: Run `fx set-device foo-bar-baz-quux`, then `fx -x shell`, observe failure to connect (device does not exist)
Test: Run `fx help set-device`, observe help documentation.
Test: Run `fx`, observe documented -d flag.
Change-Id: I078e884ee7ec0a0166d5171ce9d80d16cbba771a