[mediaplayer] add SourceImpl
This CL introduces SourceImpl, an implementation of
fuchsia.media.Source. For now, sources are not exposed to the client,
but are instead used within PlayerImpl to hold the infrastructure
relating to a source. Previously, Reader played this role, with the
Demux and DemuxSourceSegment being created just before calling
SetSourceSegment on the PlayerCore object.

In the new arrangement, demux creation, source segment creation and
source segment provisioning happen when the SetXxxSource method is
called. SourceImpl is used to hold the source segment until it's time
to call SetSourceSegment.

In future CLs, the client will be given the capability of creating
sources in advance of using them. This functionality is exposed as
the SourceManager aspect of Player. This will be the only way to
use a StreamSource, which allows clients to provide their own
elementary streams to the player.

TEST: no behavior change, CQ and manual tests for regressions
Change-Id: Ia9213689606ceef9dd9544a0c0f22624f3155bfe