[ledger] Move LevelDb initialization in LevelDbFactory.
This CL moves initialization of LevelDb from PageDb to LevelDbFactory.
It also removes PageDb::Init() method, which is not doing anything
anymore.

Test: Existing tests pass

LE-617 #comment Initialize LevelDb instances in LevelDbFactory.

This will allow in follow up CLs to pre-cache initialized instances
of LevelDb and avoid waiting for them to be created on demand.

Change-Id: I063a7606e922645b52e1198463b02639a017d83e