[inspect] Explicit null-byte handling in names and values.
We have a use case to support binary data in values, which was intended
to be supported. Unfortunately, use of c_str() prevented this from
working.

This change uses strings directly and avoids c_str() to support
null-bytes and binary data in values.

We also explicitly disallow the use of null bytes in names for metrics,
properties, and children. In debug builds it will cause an assertion
failure, but in release builds it simply returns false when attempting
to set.

TEST=run_test_component inspect_integration_tests

Change-Id: I422f625c47fad70c2616cfad34fb2e22f2d705db