[imx8][gpio] refactor gpio driver into common+soc files
The current driver is split into a file containing the imx8-family
common logic, and an imx8m-specific logic file.

This is done in preparation of adding a new driver for imx8m-mini SoC,
which shares most of the logic with the imx8m driver, though has
different iomux register options (and requires different logic).

Test: make
Change-Id: I3edc49c02e406175054faf219e9c3591b35d993c