[scenic][input] Hack: Set focus on ImportNode.
Sometimes the property is set after the view is imported.  This CL
triggers a change to the scene graph when the property is adjusted by
the client's parent.

This CL is a HACK, remove ASAP when v1 view properties migrate to Scenic.

SCN-1026 #comment

SCN-985 #comment

SCN-840 #comment

SCN-1016 #comment

TEST=manual test of full stack

Change-Id: I1ea7a75061f0ccb6474cc22c793c9bc538f28bb8