[scenic][input] Disable parallel dispatch.
The scenic input ("v2") system uses parallel dispatch as the default
policy.  This is problematic for the input v2 transition, because the
ViewManager input system implements "dispatch-to-top".

To match the behavior for transitioning, we turn off parallel dispatch
in the RootPresenter. Tests are unaffected and may continue to assume
parallel dispatch for development purposes.

One example where this CL is required is web_view. Parallel dispatch to
both IME and web_view causes focus issues.

There will be a future CL which will flip back to parallel dispatch.
This can be done after the Gesture Disambiguation protocol and framework
have been defined and implemented, and clients speak the GD protocol.

SCN-1016 #done

SCN-938 #comment

SCN-1047 #comment

TEST=CQ and manual test of full stack

Change-Id: If10ae98538e5e6f94bf58de1016dc33dd2c7e9e7