[bt][gap] Lower time to discover BR/EDR
Using the default values for Inquiry Scan Activity and Type was causing
problems with finding the device quickly when it was set discoverable.

Set the inquiry scan interval to 300 ms, leaving the inquiry scan
window at 10 ms.  This makes the device much quicker to find when
inquiry is active on another device.

NET-1742 #done
Test: Updated unit tests.  Make device discoverable and try to pair with
      it from an Android device.

Change-Id: I1474b4f2342fb6a3d038dfcbe2262469671bbc1a