[debugger] Hookep up system settings with observers
Created the detailed printing for a particular setting.

Added the SettingStoreObserver which lets someone listen for changes in
a particular setting.

Defined a setting schema definition where all options will be added.
Added the SymbolPath settting to the system and hooked it up with the
command line initializatio. Hooking up to changes still pending.

TEST=unit, manual

Change-Id: If42a37b49428b7aa3afb15df1728e1f341064ce8