[debugger] Use the frame base from DWARF.
Previously the code assumed the frame base was always the BP register
reported by the backend unwinder.

But the DWARF symbols can declare a specific frame base for a function
that should be used instead. In Rust and some optimized C++ code our
compilers like to use the SP register on x86 instead. When this happends
local variable computation would fail.

This patch adds the frame base to the Function symbol object and hooks
up the computation of this value to the Frame object.

This is complicated greatly by the fact that getting this data is now
asynchronous because arbitrary register or memory may need to be fetched
to compute the base pointer (meaning it's asynchronous) while before all
accesses were synchronous. Some callers were not updated to do the
asynchronous operation because it's messy and  currently we do not have
examples of symbols that require this. The included unit test does test
the asynchronous call, and this is handled by the "print" evaluator.

Some additional initialization was required to set the proper
architecture for the test session. This exposed a problem where InitLLVM
was called too many times. Even though there was only ever one InitLLVM
alive at a time, LLVM seems to expect this is only called once per
process. This is moved to a global singleton instead of being a member
of the ArchInfo object.

RegisterSet were made copyable since this simplified the callbacks and
there's no reason it shouldn't be copyable.

TEST=unit,manual

Change-Id: Ifd37243a47782769fb1b776227a9c9857877a486