[zxcrypt] Remove on unbind
This CL simplifies the atomic state of the zxcrypt driver, and calls
DdkRemove from DdkUnbind.  It was previously assumed that the device
should not call remove until its in-flight work was complete.  However,
DdkRelease will not be called until open connections to the device are
closed, i.e. the in-flight work is complete.  Even if this was not true,
DdkRelease will join the worker threads ensuring the operations are
complete.  As it stands now, not calling DdkUnbind leaves the inactive
device in the device tree, allowing tests to open it during rebinding
instead of waiting for the new device.

ZX-2730 #comment  This should be the necessary fix
ZX-2442 #comment  This should be the necessary fix

Tests: zxcrypt unit tests
Change-Id: Ica685d3794bd905fae40a6cf5b8976caca305f4c