[devmgr] Rework dc_watcher tracking
We now use unique_ptrs and typed intrusive lists to manage all of them.
For code clarity, I switched to a doubly linked list, because the
deletion logic for a singly linked list can be harder to follow.  This
patch contains one behavior change, which is that I believe that the
zx_channel_write failure case may leak an arbitrary number of watchers
due to mistracking of previous node pointer.  This patch adds a comment
to call out an existing use-after-free bug, which will be fixed in a
follow-up.

Test: Ran under qemu with ASAN

Change-Id: Ia11d2407577c04708a5b115bd1981499c87d93fa