[aml-thermal][vim2] Fix bug in vim2 thermal driver
This was an interesting bug.

The kernel would boot and start the aml thermal driver. The driver would
start its polling thread and poll the SCPI mailbox in a tight loop
instead of waiting 5 seconds between measurements. Eventually this would
saturate the co-processor which would become unresponsive.

The _next time_ the kernel booted there would be a chance that the
temperature measurement hw was still unhappy, this would cause the
thermal driver to hang in "aml_thermal_bind." Hanging in bind is very
bad as it causes the entire devhost to become unresponsive. This would
make any future calls to the devhost (unbind, release, suspend, etc)
fail because the devhost was still stuck in the bind call.

As a result, the symptoms appeared to be mexec failing because devmgr
would call "suspend" on the thermal driver which was still stuck in
bind. As a result, devmgr would be stuck waiting for the driver to
respond to its request to suspend forever, and it would never get around
to actually calling mexec.

Going forward we should probably have devmgr warn loudly if a driver
doesn't return from bind after a given amount of time.

Test: booted on the vim2 and observed that the thermal driver wasn't
polling in a tight loop. Also observed that I could dm reboot or mexec
in a loop

ZX-2861 #done

Change-Id: I6630add83082351bc2d6d51db842ecb167510a5c