[mediaplayer] refine payload config compatability
This CL modifies the PayloadManager so that when one connector is
providing VMOs and making no promises about how many VMOs there will
be, we ensure that the other connector is OK with this before
declaring the two configurations compatible. When the configurations
are not compatible, packets get copied.

This came up because of a new test in which the client submits an
elementary stream containing LPCM content. Because no decoder is
required, the audio renderer and the 'sink' that receives the packets
from the client are connected directly. The sink specifies
VmoAllocation::kUnrestricted, and the audio renderer specifies
VmoAllocation::kSingleVmo. Because the sink is providing VMOs, the
'unrestricted' value is interpreted to mean that no promises are made
in terms of the number of VMOs that are presented. Given that the
PayloadManager isn't equipped to deal with this situtation
dynamically (reconfiguring if more than one VMO is submitted), it
ensures compatibility by copying packets.

TEST: no new behavior, new CQ test on its way
Change-Id: I604a4bce1451733363a442a1943b2050298ca1e6