[mediaplayer] get StreamType encodings in sync with fuchsia.media
mediaplayer::StreamType contains encoding strings that are supposed
to match the encoding strings in fuchsia.media. There's some code
that verifies the values are identical. This code runs as part of a
new feature, so the problem was detected.

StreamType doesn't take a dependency on fuchsia.media, so this
agreement isn't automatic.

TEST: no new behavior
Change-Id: Ibb0a73c452bf6c717b5791affeefbf242a5a8856