[bt][l2cap] Provide unique ids for channels
When connecting from a second device to a profile, we need to create a
socket with the same channel id, but which is a different channel
because it is on a different link.  Add l2cap::Channel::unique_id()
which returns a unique id for each separate channel including the
link.

NET-1743 #done
Test: DATA_SocketFactoryTest.SameChannelDifferentHandles,
      L2CAP_ChannelTest.UniqueId

Change-Id: Ia283f70e1c0509129e816305d80a17d735fe018d