[cobalt-client]: Add MetricOptions
Move metric options into their own struct, in preparation
for adding local metrics, which can either be a local real time
version of remote metrics or be alone.

Next steps:
- Provide Local Version of Counter and Histograms (Base is enough).
- Provide options for resetting or not local metrics on query, this
would mimick iostats-like behavior.
- Provide a PseudoFile for all metrics, or a PseudoDirectory (Lazy)
with a file per metric. TBD

This is meant for instrospect API for processes that decide to expose
debug data to the local machine, and whose lifetime is tied to the
process lifetime, this metrics exist from process startup up to
the instant before process death.

TEST=cobalt-client-test

Change-Id: Ic7db1abb407290a4de0fee80f15c0d649cf5705d