[mediaplayer] flush a queue in video renderer on flush
This CL addresses an omission in the flush logic of the video
renderer in which a queue of pending packets wasn't flushed. The
result of this omission is that packets from before the flush
were presented using a timeline from after the flush, producing
retrograde presentation times, that were in turn passed to
scenic. Scenic closed the connection.

TEST: tested manually, QA test in the works

QA-452 #done

Change-Id: Ic7df4ad59f37af5d32c339b4db19899f26890a60