[inspect] Refactor values of properties into a union, remove .data.
With this change, clients can explicitly store property values as a
string or a vector of bytes.

Formatting for the purposes of JSON and text currently does not make a
distinction between these two value types, but in the future this will
be important to allow other tools to use the correct formatting.

Additionally, this removes the deprecated .data file.

TEST=run-tests

Change-Id: I1a7e235b0421053ea2fb8ea573acd54e49fb4f7d