[auth] Add a new account manager method.
Previously we only had a single method that provisioned a Fuchsia
account from a service provider account, but we're not actually able to
implement the "from a service provider account part". Since we needed to
add a provisioning method for an empty account anyway its better to do
and implement that one now rather than create unexpected behavior on
another method.

Tested: CQ unit test
Change-Id: I154232badd14dec97beddba6a43c383807423ea0