[zxio] Switch pipes to zxio codepath
This CL is a second attempt to switch pipes over to use the zxio
backend. The previous CL uncovered an issue that was not caught by
automated testing. The code has been fixed (and a test introduced) in a
previous CL. We should now be ready to reland this flag change.

Test: fdio-test
Change-Id: I52dfbdc487883d92467efa581de48633bdde8342